cloud_functions/climateiq_trigger_export_pipeline_cf/requirements.txt
affine==2.4.0
attrs==23.2.0
bcrypt==4.1.3
black==24.4.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudevents==1.10.1
contourpy==1.2.1
cryptography==42.0.6
cycler==0.12.1
deprecation==2.1.0
fiona==1.9.6
flake8==7.0.0
flake8-docstrings==1.7.0
Flask==3.0.3
fonttools==4.51.0
functions-framework==3.5.0
geopandas==0.14.4
google-api-core==2.18.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-pubsub==2.21.5
google-cloud-storage==2.16.0
google-cloud-tasks==2.16.4
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
h3==3.7.7
h3pandas==0.2.6
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
pathlib==1.0.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyproj==3.6.1
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
rasterio==1.3.10
requests==2.31.0
rsa==4.9
shapely==2.0.4
six==1.16.0
snowballstemmer==2.2.0
snuggs==1.4.7
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.2
